requirements.txt
Package                 Version
----------------------- -------
-ip                     20.2.2
asgiref                 3.2.10
beautifulsoup4          4.9.1
chardet                 3.0.4
Django                  3.1
django-ckeditor         5.9.0
django-js-asset         1.2.2
django-mdeditor         0.1.18
django-mptt             0.11.0
django-password-reset   2.0
django-simpleui         2020.9
django-tctip            1.4.2
docutils                0.16
Markdown                3.2.2
mtable                  0.1.9
Pillow                  7.2.0
pip                     20.2.3
Pygments                2.6.1
PyQt5                   5.15.1
PyQt5-sip               12.8.1
python-markdown-mathjax 1.0.4
pytz                    2020.1
QScintilla              2.11.5
setuptools              49.6.0
soupsieve               2.0.1
sqlparse                0.3.1
wcwidth                 0.2.5
